+++ src/SecucardConnect/Product/Payment/SecupaySofortService.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SecucardConnect\Product\Payment;
+
+use SecucardConnect\Product\Payment\Service\PaymentService;
+
+/**
+ * Operations for the payment.secupaysofort resource.
+ * @package SecucardConnect\Product\Payment
+ */
+class SecupaySofortService extends PaymentService
+{
+
+}